refactored test cases for selection ops
